requirements.txt
streamlit==1.36.0
openai==1.40.0
httpx<0.28

python-dotenv==1.0.1
pdfplumber==0.11.1
pypdf==4.3.1
Pillow==10.4.0

numpy==1.26.4
scikit-learn==1.5.1

# OCR-specific
pdf2image==1.17.0
pytesseract==0.3.10
